split messages per bugzilla product
Given the X-Bugzilla-Product, a hardcoded hash map the product name to a
filename. The default directory logs are written to is /var/wikibugs
which can be overridden using --dir.

The 'default' product is the fallback and will log to a new file named
mediawiki.log. The old wikibugs.log filename is no more used.

Please note that the ircecho installation on mchenry will need to be
updated to read from the new log files.

With --stdout, message will be printed to you the Product name being
prefixed and separated with a pipe '|'.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/tools/wikibugs@115410 dd0e9695-b195-4be7-bd10-2dea1a65a6b6